+++ slim/__init__.py
@@ -1,7 +1,8 @@
 from .base.app import Application, CORSOptions
+from .base.permission import ALL_PERMISSION, EMPTY_PERMISSION
 from .utils.json_ex import json_ex_dumps, json_ex_default
 from . import base
 from . import support
 from . import utils
 
-__version__ = '0.4.0 alpha'
+__version__ = '0.4.0'