+++ slim_cli/template/app.py
@@ -1,9 +1,10 @@
-from slim import Application, CORSOptions
+from slim import Application, CORSOptions, ALL_PERMISSION, EMPTY_PERMISSION
 import config
 
 
 app = Application(
     log_level=config.DEBUG,
     cookies_secret=config.COOKIE_SECRET,
+    permission=ALL_PERMISSION,
     cors_options=CORSOptions('*', allow_credentials=True, expose_headers="*", allow_headers="*")
 )